+++ src/components/database/TypenseDatabase.ts
@@ -1,5 +1,3 @@
-import fs from 'fs'
-import path from 'path'
 import { OceanNodeDBConfig } from '../../@types/OceanNode.js'
 import { convertTypesenseConfig, Typesense, TypesenseError } from './typesense.js'
 import { TypesenseSchema } from './TypesenseSchemas.js'
@@ -14,11 +12,8 @@ import {
   AbstractDdoStateDatabase,
   AbstractIndexerDatabase,
   AbstractLogDatabase,
-  AbstractNonceDatabase,
   AbstractOrderDatabase
 } from './BaseDatabase.js'
-import { SQLiteProvider } from './sqlite.js'
-import { URLUtils } from '../../utils/url.js'
 
 export class TypesenseOrderDatabase extends AbstractOrderDatabase {
   private provider: Typesense
@@ -631,150 +626,6 @@ export class TypesenseDdoDatabase extends AbstractDdoDatabase {
       }
     }
     return numDeleted
-  }
-}
-
-export class TypesenseNonceDatabase extends AbstractNonceDatabase {
-  private provider: Typesense | SQLiteProvider
-
-  constructor(config: OceanNodeDBConfig, schema: TypesenseSchema) {
-    super(config, schema)
-    return (async (): Promise<TypesenseNonceDatabase> => {
-      if (this.config.url && URLUtils.isValidUrl(this.config.url)) {
-        try {
-          this.provider = new Typesense({
-            ...convertTypesenseConfig(this.config.url),
-            logger: DATABASE_LOGGER
-          })
-          await this.provider.collections(this.schema.name).retrieve()
-        } catch (error) {
-          if (error instanceof TypesenseError && error.httpStatus === 404) {
-            await (this.provider as Typesense).collections().create(this.schema)
-          }
-        }
-      } else {
-        // Fall back to SQLite
-        DATABASE_LOGGER.logMessageWithEmoji(
-          'Typesense not available, falling back to SQLite',
-          true,
-          GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-          LOG_LEVELS_STR.LEVEL_WARN
-        )
-
-        // Ensure the directory exists before instantiating SQLiteProvider
-        const dbDir = path.dirname('databases/nonceDatabase.sqlite')
-        if (!fs.existsSync(dbDir)) {
-          fs.mkdirSync(dbDir, { recursive: true })
-        }
-        this.provider = new SQLiteProvider('databases/nonceDatabase.sqlite')
-        await this.provider.createTable()
-      }
-      return this
-    })() as unknown as TypesenseNonceDatabase
-  }
-
-  async create(address: string, nonce: number) {
-    try {
-      if (this.provider instanceof Typesense) {
-        return await this.provider
-          .collections(this.schema.name)
-          .documents()
-          .create({ id: address, nonce })
-      } else {
-        return await this.provider.create(address, nonce)
-      }
-    } catch (error) {
-      const errorMsg =
-        `Error when creating new nonce entry ${nonce} for address ${address}: ` +
-        error.message
-      DATABASE_LOGGER.logMessageWithEmoji(
-        errorMsg,
-        true,
-        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-        LOG_LEVELS_STR.LEVEL_ERROR
-      )
-      return null
-    }
-  }
-
-  async retrieve(address: string) {
-    try {
-      if (this.provider instanceof Typesense) {
-        return await this.provider
-          .collections(this.schema.name)
-          .documents()
-          .retrieve(address)
-      } else {
-        return await this.provider.retrieve(address)
-      }
-    } catch (error) {
-      const errorMsg =
-        `Error when retrieving nonce entry for address ${address}: ` + error.message
-      DATABASE_LOGGER.logMessageWithEmoji(
-        errorMsg,
-        true,
-        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-        LOG_LEVELS_STR.LEVEL_ERROR
-      )
-      return null
-    }
-  }
-
-  async update(address: string, nonce: number) {
-    try {
-      if (this.provider instanceof Typesense) {
-        return await this.provider
-          .collections(this.schema.name)
-          .documents()
-          .update(address, { nonce })
-      } else {
-        return await this.provider.update(address, nonce)
-      }
-    } catch (error) {
-      if (
-        this.provider instanceof Typesense &&
-        error instanceof TypesenseError &&
-        error.httpStatus === 404
-      ) {
-        return await this.provider
-          .collections(this.schema.name)
-          .documents()
-          .create({ id: address, nonce })
-      }
-      const errorMsg =
-        `Error when updating nonce entry ${nonce} for address ${address}: ` +
-        error.message
-      DATABASE_LOGGER.logMessageWithEmoji(
-        errorMsg,
-        true,
-        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-        LOG_LEVELS_STR.LEVEL_ERROR
-      )
-      return null
-    }
-  }
-
-  async delete(address: string) {
-    try {
-      if (this.provider instanceof Typesense) {
-        return await this.provider
-          .collections(this.schema.name)
-          .documents()
-          .delete(address)
-      } else {
-        return await this.provider.delete(address)
-      }
-    } catch (error) {
-      const errorMsg =
-        `Error when deleting nonce entry for address ${address}: ` + error.message
-      DATABASE_LOGGER.logMessageWithEmoji(
-        errorMsg,
-        true,
-        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-        LOG_LEVELS_STR.LEVEL_ERROR
-      )
-      return null
-    }
   }
 }
