+++ src/test/unit/indexer/indexer.test.ts
@@ -4,6 +4,10 @@ import { OceanIndexer } from '../../../components/Indexer/index.js'
 import { getConfiguration } from '../../../utils/index.js'
 import { Database } from '../../../components/database/index.js'
 import { OceanNodeConfig } from '../../../@types/OceanNode.js'
+import {
+  hasValidDBConfiguration,
+  isReachableConnection
+} from '../../../utils/database.js'
 
 describe('OceanIndexer', () => {
   let oceanIndexer: OceanIndexer
@@ -20,8 +24,18 @@ describe('OceanIndexer', () => {
     expect(oceanIndexer.getDatabase().getConfig()).to.be.equal(mockDatabase.getConfig())
     expect(oceanIndexer.getIndexingQueue().length).to.be.equal(0)
     expect(oceanIndexer.getJobsPool().length).to.be.equal(0)
-    // cannot start threads without DB connection
-    expect(await oceanIndexer.startThreads()).to.be.equal(false)
+
+    if (
+      hasValidDBConfiguration(mockDatabase.getConfig()) &&
+      (await isReachableConnection(mockDatabase.getConfig().url))
+    ) {
+      // should be fine, if we have a valid RPC as well
+      expect(await oceanIndexer.startThreads()).to.be.equal(true)
+    } else {
+      // cannot start threads without DB connection
+      expect(await oceanIndexer.startThreads()).to.be.equal(false)
+    }
+
     // there are no worker threads available
     expect(oceanIndexer.stopAllThreads()).to.be.equal(false)
   })
